Do not cache database properties dialog
